Performance optimization. does not create an EventHandlers object unless absolutely needed. In theory the ram gains should be bigger than the cpu gains.

```
after

-----------
Small Model
-----------
      31ms |      x 100 |  0.3ms avg
     182ms |    x 1,000 |  0.2ms avg
     794ms |   x 10,000 |  0.1ms avg
     163ms |    x 1,000 |  0.2ms avg
      15ms |      x 100 |  0.1ms avg

------------
Medium Model
------------
      34ms |      x 100 |  0.3ms avg
     325ms |    x 1,000 |  0.3ms avg
   1,661ms |   x 10,000 |  0.2ms avg
     295ms |    x 1,000 |  0.3ms avg
      43ms |      x 100 |  0.4ms avg

------------------------
Large Model - 0 children
------------------------
     110ms |      x 100 |  1.1ms avg
     665ms |    x 1,000 |  0.7ms avg
      97ms |      x 100 |  1.0ms avg

-------------------------------------------
Large Model - 10 small & 10 medium children
-------------------------------------------
      99ms |       x 50 |  2.0ms avg
     336ms |      x 250 |  1.3ms avg
      98ms |       x 50 |  2.0ms avg

-------------------------------------------
Large Model - 100 small & 0 medium children
-------------------------------------------
     113ms |       x 50 |  2.3ms avg
     314ms |      x 250 |  1.3ms avg
      51ms |       x 50 |  1.0ms avg

-------------------------------------------
Large Model - 0 small & 100 medium children
-------------------------------------------
     342ms |       x 50 |  6.8ms avg
   1,267ms |      x 250 |  5.1ms avg
     325ms |       x 50 |  6.5ms avg

20.60user 4.18system 0:11.89elapsed 208%CPU (0avgtext+0avgdata 264200maxresident)k
0inputs+0outputs (0major+197978minor)pagefaults 0swaps
Done in 42.04s.

before

-----------
Small Model
-----------
      34ms |      x 100 |  0.3ms avg
     177ms |    x 1,000 |  0.2ms avg
     838ms |   x 10,000 |  0.1ms avg
     170ms |    x 1,000 |  0.2ms avg
      12ms |      x 100 |  0.1ms avg

------------
Medium Model
------------
      41ms |      x 100 |  0.4ms avg
     363ms |    x 1,000 |  0.4ms avg
   1,804ms |   x 10,000 |  0.2ms avg
     308ms |    x 1,000 |  0.3ms avg
      37ms |      x 100 |  0.4ms avg

------------------------
Large Model - 0 children
------------------------
     114ms |      x 100 |  1.1ms avg
     696ms |    x 1,000 |  0.7ms avg
      83ms |      x 100 |  0.8ms avg

-------------------------------------------
Large Model - 10 small & 10 medium children
-------------------------------------------
     108ms |       x 50 |  2.2ms avg
     365ms |      x 250 |  1.5ms avg
     119ms |       x 50 |  2.4ms avg

-------------------------------------------
Large Model - 100 small & 0 medium children
-------------------------------------------
      88ms |       x 50 |  1.8ms avg
     343ms |      x 250 |  1.4ms avg
      90ms |       x 50 |  1.8ms avg

-------------------------------------------
Large Model - 0 small & 100 medium children
-------------------------------------------
     358ms |       x 50 |  7.2ms avg
   1,384ms |      x 250 |  5.5ms avg
     328ms |       x 50 |  6.6ms avg

21.26user 5.35system 0:12.59elapsed 211%CPU (0avgtext+0avgdata 260288maxresident)k
0inputs+0outputs (0major+206645minor)pagefaults 0swaps
Done in 43.47s.
```