@xaviergonz 2019-01-30
@k-g-a would you be so kind as to test it against your playground that you showed in https://github.com/mobxjs/mobx-state-tree/issues/1151#issuecomment-458468637 ?