@k-g-a 2019-01-31
Made a quick glance: 550Kb out of 7.8Mb are wiped out, thats around 7%. 
![image](https://user-images.githubusercontent.com/12710451/52078115-4399ba80-25b4-11e9-83a1-63e0829d4dd7.png)
![image](https://user-images.githubusercontent.com/12710451/52078153-590ee480-25b4-11e9-87dd-a0bdcd2fde88.png)


That's a good cut as many users payed for nothing. 
But if I understand correclty, should I use safe reference for my GrandChild (deepest nodes) - those Kbs will come back?