@xaviergonz 2019-01-31
> But if I understand correclty, should I use safe reference for my GrandChild (deepest nodes) - those Kbs will come back?

Yep, if a hook is attached (such as what happens when using safe reference) then those KBs will come back, but I don't see many ways around that one :)